Re-factor of container cmd functions to seperate file
